Pojo: Don't throw if field type != property type
If a property was discovered through reflection, we shouldn't be heavily
introspecting the underlying property and enforcing that the type
matches. How the object stores the field internally should be treated as
an implementation detail if proper getters/setters/constructors are provided.

This change is made so it remains backwards compatible, in that things
that successfully serialized using a combination of properties and
direct field access will still serialize/deserialize.